Fjerne mer gammel felleskode
