+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/Brukernotifikasjon.kt
@@ -1,4 +1,4 @@
-package no.nav.personbruker.dittnav.eventaggregator.common.database.entity
+package no.nav.personbruker.dittnav.eventaggregator.common
 
 import no.nav.personbruker.dittnav.eventaggregator.config.EventType
 import no.nav.personbruker.dittnav.eventaggregator.done.Done
