--- src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/database/BrukernotifikasjonRepository.kt
@@ -1,9 +1,0 @@
-package no.nav.personbruker.dittnav.eventaggregator.common.database
-
-interface BrukernotifikasjonRepository<T> {
-
-    suspend fun createInOneBatch(entities: List<T>): ListPersistActionResult<T>
-
-    suspend fun createOneByOneToFilterOutTheProblematicEvents(entities: List<T>): ListPersistActionResult<T>
-
-}
