+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/done/DonePersistingService.kt
@@ -1,7 +1,7 @@
 package no.nav.personbruker.dittnav.eventaggregator.done
 
 import no.nav.personbruker.dittnav.eventaggregator.common.LocalDateTimeHelper.nowAtUtc
-import no.nav.personbruker.dittnav.eventaggregator.common.database.entity.Brukernotifikasjon
+import no.nav.personbruker.dittnav.eventaggregator.common.Brukernotifikasjon
 
 class DonePersistingService(private val doneRepository: DoneRepository) {
