+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/done/doneQueries.kt
@@ -1,8 +1,10 @@
 package no.nav.personbruker.dittnav.eventaggregator.done
 
+import no.nav.personbruker.dittnav.eventaggregator.common.Brukernotifikasjon
 import no.nav.personbruker.dittnav.eventaggregator.common.database.ListPersistActionResult
 import no.nav.personbruker.dittnav.eventaggregator.common.database.PersistActionResult
 import no.nav.personbruker.dittnav.eventaggregator.common.database.util.*
+import no.nav.personbruker.dittnav.eventaggregator.config.EventType
 import java.sql.Connection
 import java.sql.PreparedStatement
 import java.sql.ResultSet
@@ -78,6 +80,25 @@ fun ResultSet.toDoneEvent(): Done {
     )
 }
 
+
+fun Connection.getBrukernotifikasjonFromViewForEventIds(eventIds: List<String>): List<Brukernotifikasjon> =
+    prepareStatement("""SELECT brukernotifikasjon_view.* FROM brukernotifikasjon_view WHERE eventid = ANY(?)""")
+        .use {
+            it.setArray(1, toVarcharArray(eventIds))
+            it.executeQuery().list {
+                toBrukernotifikasjon()
+            }
+        }
+
+private fun ResultSet.toBrukernotifikasjon(): Brukernotifikasjon {
+    return Brukernotifikasjon(
+        eventId = getString("eventId"),
+        systembruker = getString("systembruker"),
+        type = EventType.valueOf("${getString("type")}_intern".uppercase()),
+        fodselsnummer = getString("fodselsnummer")
+    )
+}
+
 private fun PreparedStatement.buildStatementForSingleRow(doneEvent: Done) {
     setString(1, doneEvent.systembruker)
     setObject(2, doneEvent.eventTidspunkt, Types.TIMESTAMP)
@@ -89,3 +110,5 @@ private fun PreparedStatement.buildStatementForSingleRow(doneEvent: Done) {
     setString(8, doneEvent.appnavn)
     setObject(9, doneEvent.sistBehandlet, Types.TIMESTAMP)
 }
+
+
