+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/metrics/PrometheusMetricsCollector.kt
@@ -15,10 +15,7 @@ object PrometheusMetricsCollector {
             .labelNames("type", "producer")
             .register()
 
-
     fun registerEventsCached(count: Int, eventType: EventType, producer: String) {
         MESSAGES_CACHED.labels(eventType.eventType, producer).inc(count.toDouble())
     }
-
-
 }
