+++ src/test/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/objectmother/BrukernotifikasjonObjectMother.kt
@@ -1,7 +1,7 @@
 package no.nav.personbruker.dittnav.eventaggregator.common.objectmother
 
 import no.nav.personbruker.dittnav.eventaggregator.beskjed.Beskjed
-import no.nav.personbruker.dittnav.eventaggregator.common.database.entity.Brukernotifikasjon
+import no.nav.personbruker.dittnav.eventaggregator.common.Brukernotifikasjon
 import no.nav.personbruker.dittnav.eventaggregator.config.EventType
 import no.nav.personbruker.dittnav.eventaggregator.innboks.Innboks
 import no.nav.personbruker.dittnav.eventaggregator.oppgave.Oppgave
